+++ NEWS.md
@@ -1,3 +1,11 @@
+Changes in version 1.4
+
+Bug fixes
+
++ (v. 1.4.1) `projects()` constructs the complete tibble of projects
+  by paging through result sets of size 100, to accommodate upstream
+  changes. See https://github.com/Bioconductor/hca/issues/33
+
 Changes in version 1.2
 
 New features